refactor: improve options typing
