+++ src/app/core/database/database.ts
@@ -15,8 +15,6 @@
  *     along with ndb-core.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { QueryOptions } from "./query-options";
-
 /**
  * An implementation of this abstract class provides functions for direct database access.
  * This interface is an extension of the [PouchDB API](https://pouchdb.com/api.html).
@@ -30,7 +28,7 @@ export abstract class Database {
    * @param id The primary key of the document to be loaded
    * @param options Optional options for the database engine (PouchDB)
    */
-  abstract get(id: string, options?: QueryOptions): Promise<any>;
+  abstract get(id: string, options?: GetOptions): Promise<any>;
 
   /**
    * Load all documents (matching the given PouchDB options) from the database.
@@ -41,7 +39,7 @@ export abstract class Database {
    *
    * @param options PouchDB options object as in the normal PouchDB library
    */
-  abstract allDocs(options?: QueryOptions): Promise<any>;
+  abstract allDocs(options?: GetAllOptions): Promise<any>;
 
   /**
    * Save a document to the database.
@@ -88,3 +86,28 @@ export abstract class Database {
     });
   }
 }
+
+/**
+ * Basic query options supported by {@link Database}.
+ *
+ * also see https://pouchdb.com/guides/queries.html
+ */
+export type QueryOptions = PouchDB.Query.Options<any, any>;
+
+/**
+ * Basic database read options supported by {@link Database}.
+ *
+ * also see https://pouchdb.com/api.html#fetch_document
+ */
+export type GetAllOptions =
+  | PouchDB.Core.AllDocsWithKeyOptions
+  | PouchDB.Core.AllDocsWithKeysOptions
+  | PouchDB.Core.AllDocsWithinRangeOptions
+  | PouchDB.Core.AllDocsOptions;
+
+/**
+ * Basic database read options supported by {@link Database}.
+ *
+ * also see https://pouchdb.com/api.html#fetch_document
+ */
+export type GetOptions = PouchDB.Core.GetOptions;
