+++ src/app/core/database/mock-database.ts
@@ -15,11 +15,10 @@
  *     along with ndb-core.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { Database } from "./database";
+import { Database, GetAllOptions, QueryOptions } from "./database";
 import { Note } from "../../child-dev-project/notes/model/note";
 import { AttendanceMonth } from "../../child-dev-project/attendance/model/attendance-month";
 import { ChildSchoolRelation } from "../../child-dev-project/children/model/childSchoolRelation";
-import { QueryOptions } from "./query-options";
 
 /**
  * In-Memory database implementation that works as a drop-in replacement of {@link PouchDatabase}
@@ -62,12 +61,12 @@ export class MockDatabase extends Database {
    * see {@link Database}
    * @param options Only 'startkey' is considered by the MockDatabase implementation
    */
-  allDocs(options?: QueryOptions) {
+  allDocs(options?: GetAllOptions) {
     let result = this.data;
 
     // default options provided through getAll(prefix): {include_docs: true, startkey: prefix, endkey: prefix + '\ufff0'}
     // MockDatabase ignores endkey and only implements filtering based on startkey/prefix
-    if (options && options.hasOwnProperty("startkey")) {
+    if (options && "startkey" in options) {
       result = this.data.filter((o) => o._id.startsWith(options.startkey));
     }
